Use subpixel precision for source in the paint functions of all backends

+++ src/openslide-ops-jpeg.c
@@ -2,6 +2,7 @@
  *  OpenSlide, a library for reading whole slide image files
  *
  *  Copyright (c) 2007-2009 Carnegie Mellon University
+ *  Copyright (c) 2011 Google, Inc.
  *  All rights reserved.
  *
  *  OpenSlide is free software: you can redistribute it and/or modify
@@ -748,8 +749,8 @@ static void paint_region(openslide_t *osr, cairo_t *cr,
 
   // compute coordinates
   double ds = openslide_get_layer_downsample(osr, layer);
-  int64_t ds_x = x / ds;
-  int64_t ds_y = y / ds;
+  double ds_x = x / ds;
+  double ds_y = y / ds;
   int64_t start_tile_x = ds_x / l->tile_advance_x;
   double offset_x = ds_x - (start_tile_x * l->tile_advance_x);
   int64_t end_tile_x = ((ds_x + w) / l->tile_advance_x) + 1;
