+++ src/openslide-ops-ngr.c
@@ -2,6 +2,7 @@
  *  OpenSlide, a library for reading whole slide image files
  *
  *  Copyright (c) 2007-2010 Carnegie Mellon University
+ *  Copyright (c) 2011 Google, Inc.
  *  All rights reserved.
  *
  *  OpenSlide is free software: you can redistribute it and/or modify
@@ -170,14 +171,14 @@ static void paint_region(openslide_t *osr, cairo_t *cr,
 
   // compute coordinates
   double ds = openslide_get_layer_downsample(osr, layer);
-  int64_t ds_x = x / ds;
-  int64_t ds_y = y / ds;
+  double ds_x = x / ds;
+  double ds_y = y / ds;
   int64_t start_tile_x = ds_x / ngr->column_width;
   int64_t end_tile_x = ((ds_x + w) / ngr->column_width) + 1;
   int64_t start_tile_y = ds_y;
   int64_t end_tile_y = ds_y + h + 1;
 
-  int64_t offset_x = ds_x - (start_tile_x * ngr->column_width);
+  double offset_x = ds_x - (start_tile_x * ngr->column_width);
 
   _openslide_read_tiles(cr,
 			layer,
